dividing Types.Internal module.
